Move example specs
