+++ examples/express/zod/express.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { asAsync } from "../../../src/express";
-import { pathMap } from "./spec";
+import { pathMap } from "../../spec/zod";
 import { ToHandlers, typed } from "../../../src/express/zod";
 
 const emptyMiddleware = (
